requirements.txt
# HablaDB - Chat-to-SQL
streamlit>=1.28.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
duckdb-engine>=0.11.0
litellm>=1.30.0
python-dotenv>=1.0.0
pandas>=2.0.0
